+++ database/migrations/2024_10_15_183132_yatirimci_sayisi.php
@@ -4,33 +4,27 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::create('fonprices', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->string('name');
-            $table->unsignedBigInteger('fon_id')
+        if (!Schema::hasTable('yatirimciSayisi'))
+            Schema::create('yatirimciSayisi', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->unsignedBigInteger('fon_id')
                     ->foreign('fon_id')->references('id')->on('fons')
                     ->onDelete('cascade');
-            $table->float('price')->min(0);
-            $table->date('date');
-            $table->foreign('name')->references('code')->on('fons')->onDelete('cascade');
-    });
-
-
-
+                $table->integer('yatirimciSayisi')->min(0);
+                $table->date('date');
+            });
     }
-
     /**
      * Reverse the migrations.
      */
     public function down(): void
     {
-        //
+       
     }
 };
